move github to the correct folder
