+++ server/src/external/mod.rs
@@ -1,2 +1,3 @@
 pub mod connectum;
+pub mod github;
 pub mod meilisearch;
