+++ server/src/feedback/mod.rs
@@ -1,4 +1,3 @@
-mod github;
 pub mod post_feedback;
 pub mod proposed_edits;
 pub mod tokens;
